+++ pkg/util/metric/m_register.go
@@ -21,7 +21,7 @@ const constTenantKey = "account"
 // this constant lable is used for sys_* and process_* table
 var sysTenantID = prom.Labels{constTenantKey: "sys"}
 
-var initCollectors = []Collector{
+var InitCollectors = []Collector{
 	// sql metric
 	StatementCounterFactory,
 	StatementErrorsFactory,
@@ -36,27 +36,21 @@ var initCollectors = []Collector{
 	hardwareStatsCollector,
 }
 
-// register all defined collector here
-func registerAllMetrics() {
-	for _, c := range initCollectors {
-		mustRegister(c)
-	}
-}
-
 type SubSystem struct {
 	Name              string
 	Comment           string
 	SupportUserAccess bool
 }
 
-var SubSystemSql = &SubSystem{"sql", "base on query action", true}
-var SubSystemServer = &SubSystem{"server", "MO Server status, observe from inside", true}
-var SubSystemProcess = &SubSystem{"process", "MO process status", false}
-var SubSystemSys = &SubSystem{"sys", "OS status", false}
+var AllSubSystem = map[string]*SubSystem{}
 
-var AllSubSystem = map[string]*SubSystem{
-	SubSystemSql.Name:     SubSystemSql,
-	SubSystemServer.Name:  SubSystemServer,
-	SubSystemProcess.Name: SubSystemProcess,
-	SubSystemSys.Name:     SubSystemSys,
+func registerSubSystem(s *SubSystem) {
+	AllSubSystem[s.Name] = s
+}
+
+func init() {
+	registerSubSystem(&SubSystem{"sql", "base on query action", true})
+	registerSubSystem(&SubSystem{"server", "MO Server status, observe from inside", true})
+	registerSubSystem(&SubSystem{"process", "MO process status", false})
+	registerSubSystem(&SubSystem{"sys", "OS status", false})
 }
