+++ pkg/util/metric/mometric/testutil_test.go
@@ -15,11 +15,12 @@
 // this file contains test utils. Name this file "*_test.go" to make
 // compiler ignore it
 
-package metric
+package mometric
 
 import (
 	"context"
 	"sync"
+	"sync/atomic"
 	"time"
 
 	"github.com/matrixorigin/matrixone/pkg/common/moerr"
@@ -48,6 +49,13 @@ func waitWgTimeout(wg *sync.WaitGroup, after time.Duration) error {
 		return moerr.NewInternalError(context.Background(), "timeout")
 	case <-c:
 		return nil
+	}
+}
+
+func makeDummyClock(startOffset int64) func() int64 {
+	var tick = startOffset - 1
+	return func() int64 {
+		return atomic.AddInt64(&tick, 1)
 	}
 }
